tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/006.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: document.write inline - multiple</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>

	<script>log('inline script #1');
		document.write( '<script>log(\'doc write script 1\')<\/script>' );
		document.write( '<script>log(\'doc write script 2\')<\/script>' );
		eval('log(\'eval 1\')');
		document.write( '<script>log(\'doc write script 3\')<\/script>' );
		log('end script #1');
	</script>

	<script type="text/javascript">
	var t = async_test()

	function test() {
		assert_array_equals(eventOrder, ['inline script #1', 'doc write script 1','doc write script 2', 'eval 1','doc write script 3', 'end script #1' ]);
		t.done();
}
	onload = t.step_func(test)
	</script>

</body></html>
